avoid signed/unsigned warning
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4853 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/io/file/dataoutput.hh
@@ -239,7 +239,8 @@ protected:
 
       CachingQuadrature<GridPartType,0> quad(en_,df->space().order());
       LocalFunctionType lf = df->localFunction(en_);
-      for (int i=0;i<quad.nop();++i) {
+      for( size_t i = 0; i < quad.nop(); ++i )
+      {
         RangeType u;
         DomainType x = en_.geometry().global(quad.point(i));
         lf.evaluate(quad[i],u);
